pyproject.toml
[tool.poetry]
name = "led"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
viam-sdk = "0.5.1"
mypy = "1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
